Add new Ui, Storage, Parser and TaskList classes
