+++ src/main/java/fleur/Ui.java
@@ -1,0 +1,10 @@
+package fleur;
+
+public class Ui {
+
+    protected void welcomeMessage() {
+        System.out.println("'Allo! I'm Fleur.");
+        System.out.println("Tell me what you need to do, s'il vous plaît.");
+    }
+
+}
